esp32, async websocket webserver

+++ ntp.ino
@@ -5,7 +5,7 @@
 WiFiUDP ntpUDP;
 NTPClient timeClient(ntpUDP, NTP_SERVER, 0 * 3600, 60000 * 10); // update every 10 min
 
-long long EpochTime() {
+long EpochTime() {
   return timeClient.getEpochTime();
 }
